// src/app/core/models/comunidad/Comunidad.ts
export class Comunidad {
    id!: number;
    clave!: string;
    nombre!: string;
    municipio?: string;
    estado?: string;
    direccion?: string;
    telefono?: string;
    email?: string;
    codigoPostal?: string;
    descripcion?: string;
    activo?: boolean;
    createdAt?: string;
    updatedAt?: string;
}